requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
loguru==0.6.0
multidict==6.0.2
PyYAML==6.0
SQLAlchemy==1.4.40
yarl==1.8.1
